+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [4.1.3] - 2025-05-23
+- Updated to work with pymongo v4.X
+- Updated python-lib to v4.0.1
+- Updated mongo-lib to v4.5.2
+
+### Changes
+- dbcc, defrag: Replaced get_db_tbl call with mongo_libs.get_db_tbl call.
+- Documentation changes.
+
+### Removed
+- get_all_dbs_tbls function
+- get_db_tbl function
+
+
 ## [4.1.2] - 2025-03-18
 
 ### Fixed